@@ -9,15 +9,7 @@
         "dev"
     ],
     "resolutions": {
-        "@react-three/fiber": "7.0.27",
-        "@types/react": "^17.0.39",
-        "@types/react-dom": "^17.0.11",
-        "typescript": "^4.2.3",
-        "jest": "^24.8.0",
-        "react": "17.0.2",
-        "react-dom": "17.0.2",
-        "webpack": "^5.27.2",
-        "three": ">=0.133"
+        "webpack": "^5.27.2"
     },
     "scripts": {
         "build": "turbo run build",
@@ -37,44 +29,44 @@
     },
     "devDependencies": {
         "@cypress/webpack-preprocessor": "^4.1.0",
-        "@rollup/plugin-node-resolve": "^8.0.0",
+        "@rollup/plugin-node-resolve": "^13.3.0",
         "@rollup/plugin-replace": "^2.3.2",
-        "@testing-library/dom": "^6.10.1",
-        "@testing-library/react": "13.0.0-alpha.5",
-        "@testing-library/react-hooks": "8.0.0-alpha.1",
-        "@types/jest": "^23.3.9",
-        "@types/node": "^17.0.8",
-        "@types/react": "^17.0.39",
-        "@types/react-dom": "^17.0.11",
-        "@types/styled-components": "^4.1.0",
-        "@types/three": "^0.136.1",
+        "@testing-library/dom": "^8.16.0",
+        "@testing-library/jest-dom": "^5.16.4",
+        "@testing-library/react": "^13.3.0",
+        "@testing-library/react-hooks": "^8.0.1",
+        "@types/jest": "^28.1.6",
+        "@types/node": "^18.6.3",
+        "@types/react": "^18.0.15",
+        "@types/react-dom": "^18.0.6",
+        "@types/styled-components": "^5.1.25",
+        "@types/three": "^0.143.0",
         "@typescript-eslint/eslint-plugin": "^5.9.1",
         "@typescript-eslint/parser": "^5.9.1",
-        "bundlesize": "^0.18.0",
-        "concurrently": "^7.0.0",
+        "bundlesize": "^0.18.1",
+        "concurrently": "^7.3.0",
         "convert-tsconfig-paths-to-webpack-aliases": "^0.9.2",
         "cypress": "^3.4.0",
-        "eslint": "^8.6.0",
-        "eslint-config-prettier": "^8.3.0",
-        "eslint-plugin-import": "^2.25.4",
-        "eslint-plugin-react": "^7.28.0",
-        "eslint-plugin-react-hooks": "^4.3.0",
-        "jest": "^27.4.7",
-        "jest-dom": "^3.5.0",
-        "jest-junit": "^6.4.0",
-        "jest-watch-typeahead": "^0.4.2",
+        "eslint": "^8.21.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-import": "^2.26.0",
+        "eslint-plugin-react": "^7.30.1",
+        "eslint-plugin-react-hooks": "^4.6.0",
+        "jest": "^28.0.0",
+        "jest-environment-jsdom": "^28.1.3",
+        "jest-watch-typeahead": "^2.0.0",
         "lerna": "^4.0.0",
         "lint-staged": "^8.0.4",
         "prettier": "^2.5.1",
-        "react": "^17.0.2",
-        "react-dev-utils": "^11.0.4",
-        "react-dom": "^17.0.2",
+        "react": "^18.2.0",
+        "react-dev-utils": "^12.0.1",
+        "react-dom": "^18.2.0",
         "rollup": "^2.63.0",
         "rollup-plugin-dts": "^4.2.2",
         "rollup-plugin-terser": "^5.3.0",
         "rollup-plugin-visualizer": "^2.4.3",
-        "ts-jest": "^24.0.2",
-        "ts-loader": "^5.3.0",
+        "ts-jest": "^28.0.7",
+        "ts-loader": "^9.3.1",
         "turbo": "latest",
         "type-fest": "^1.0.1",
         "typescript": "^4.2.3",
