@@ -1,6 +1,6 @@
 {
     "name": "framer-motion-3d",
-    "version": "6.5.1",
+    "version": "7.0.0",
     "description": "A simple and powerful React animation library for @react-three/fiber",
     "main": "dist/cjs/index.js",
     "module": "dist/es/index.mjs",
@@ -48,45 +48,18 @@
         "postpublish": "git push --tags"
     },
     "dependencies": {
-        "framer-motion": "^6.5.1",
-        "react-merge-refs": "^1.1.0"
+        "framer-motion": "^7.0.0",
+        "react-merge-refs": "^2.0.1"
     },
     "peerDependencies": {
-        "@react-three/fiber": "^7.0.27",
-        "react": ">=16.8 || ^17.0.0",
-        "react-dom": ">=16.8 || ^17.0.0",
+        "@react-three/fiber": "^8.2.2",
+        "react": ">=18.0",
+        "react-dom": ">=18.0",
         "three": ">=0.133"
     },
     "devDependencies": {
-        "@react-three/fiber": "7.0.24",
-        "@react-three/test-renderer": "7.0.25",
-        "@rollup/plugin-commonjs": "^21.0.1",
-        "@rollup/plugin-node-resolve": "^8.0.0",
-        "@rollup/plugin-replace": "^2.3.2",
-        "@testing-library/dom": "^6.10.1",
-        "@testing-library/react": "^9.3.2",
-        "@testing-library/react-hooks": "^5.1.0",
-        "@types/react": "^17.0.39",
-        "@types/react-dom": "^17.0.11",
-        "bundlesize": "^0.18.0",
-        "convert-tsconfig-paths-to-webpack-aliases": "^0.9.2",
-        "jest": "^24.8.0",
-        "jest-dom": "^3.5.0",
-        "jest-junit": "^6.4.0",
-        "jest-watch-typeahead": "^0.4.2",
-        "prettier": "^2.1.0",
-        "react-dev-utils": "^11.0.4",
-        "rollup": "^2.10.5",
-        "rollup-plugin-terser": "^5.3.0",
-        "rollup-plugin-visualizer": "^2.4.3",
-        "styled-components": "^4.1.1",
-        "three": "^0.133.0",
-        "ts-jest": "^24.0.2",
-        "ts-loader": "^5.3.0",
-        "type-fest": "^1.0.1",
-        "typescript": "^4.2.3",
-        "webpack": "^5.27.2",
-        "yarn-deduplicate": "^1.1.1"
-    },
-    "gitHead": "c55efb119209ecad33392bb868cc6af2d38a9f40"
+        "@react-three/fiber": "^8.2.2",
+        "@react-three/test-renderer": "^9.0.0",
+        "@rollup/plugin-commonjs": "^22.0.1"
+    }
 }
