@@ -1,6 +1,6 @@
 {
     "name": "framer-motion",
-    "version": "6.5.1",
+    "version": "7.0.0",
     "description": "A simple and powerful React animation library",
     "main": "dist/cjs/index.js",
     "module": "dist/es/index.mjs",
@@ -41,7 +41,7 @@
         "clean": "rm -rf types dist lib",
         "test": "yarn test-server && yarn test-client",
         "test-ci": "yarn test",
-        "test-client": "jest --coverage --config jest.config.json --max-workers=2",
+        "test-client": "jest --config jest.config.json --max-workers=2",
         "test-server": "jest --config jest.config.ssr.json",
         "test-watch": "jest --watch --coverage --coverageReporters=lcov --config jest.config.json",
         "test-projection": "yarn run collect-projection-tests && start-server-and-test 'pushd ../../; python -m SimpleHTTPServer; popd' http://0.0.0.0:8000 'cypress run -s cypress/integration/projection.chrome.ts --config baseUrl=http://localhost:8000/'",
@@ -58,21 +58,16 @@
         "start-dev-server": "webpack serve --config ../../dev/webpack/config.js --hot"
     },
     "peerDependencies": {
-        "react": ">=16.8 || ^17.0.0 || ^18.0.0",
-        "react-dom": ">=16.8 || ^17.0.0 || ^18.0.0"
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0"
     },
     "dependencies": {
-        "@motionone/dom": "10.12.0",
+        "@motionone/dom": "10.13.1",
         "framesync": "6.0.1",
         "hey-listen": "^1.0.8",
         "popmotion": "11.0.3",
         "style-value-types": "5.1.0",
         "tslib": "^2.1.0"
-    },
-    "devDependencies": {
-        "jest": "^27.4.7",
-        "jest-dom": "^3.5.0",
-        "styled-components": "^4.1.1"
     },
     "optionalDependencies": {
         "@emotion/is-prop-valid": "^0.8.2"
@@ -102,6 +97,5 @@
             "path": "./dist/size-webpack-dom-max.js",
             "maxSize": "31.7 kB"
         }
-    ],
-    "gitHead": "c55efb119209ecad33392bb868cc6af2d38a9f40"
+    ]
 }
