@@ -1,4 +1,4 @@
-import "jest-dom/extend-expect"
+import "@testing-library/jest-dom"
 // Get fireEvent from the native testing library
 // because @testing-library/react one switches out mouseEnter and mouseLeave
 import { fireEvent, getByTestId } from "@testing-library/dom"
